Use collection object
The two post-harvest tasks now use the collection object that is passed
to them to access the metadata, rather than as a string to access the
catalog_for_provider().

I also renamed post_havest_task to post_harvest_task, and fixed what
looked like indentation errors in the yale.yaml.