Allow DAGs to be discoverable
airflow has an amusing heuristic to determine whether a folder contains
dags or not:

https://github.com/apache/airflow/blob/main/airflow/utils/file.py#L337-L357

Basically the .py file must contain the words 'airflow' and 'dag' and
since dlme_airflow.dags.harvest_catalog.py no longer does it isn't
loading it at all :)

The other problem is that the DAG classes are being created in the
dlme_airflow.services.harvest_dag_generator module instead of
dlme_airflow.dags.harvest_catalog. So even after aiflow has loaded the
file it doesn't find any DAG classes in it, and so it decides there
aren't any.

So this commit adds a comment with 'airflow' and 'dag' in it to
dlme_airflow.dags.harvest_catalog and also adjusts create_provider_dags
to take an optional module name to create the classes in.

Closes #193