XML Driver fixes
This commit addresses some things that the bnf.yaml made clear. The
record_selector was being defined in the intake catalog entries but it
wasn't actually being used by the XML Driver. Once it was being used it
needs to start with a `.` so for example `.//item`, because it is
searching on the root element at first. This isn't needed in subsequent
XPaths.

I updated the bnf entries but these should probably be checked. I mostly
wanted to test that the XPaths were working correctly.