Implement wait option for OaiXmlSource
This adds a wait option for the OaiXmlSource driver which will inject a
sleep of n seconds between sending HTTP requests to the OAI server.

Sickle allows you to pass in your own iterator class, which gets used
internally when making requests. This class needs to change its behavior
depending on the desired wait time, so it is created dynamically by the
make_iterator() function.

The QNL intake catalog entry has been adjusted to use `wait: 2` which
fixes #200.

I also added a `--log` option to bin/get to write the sleep messages (and
other diagnostics) to a file.

Closes #235