Improve multiple field handling in XmlSource driver
This is a significant refactor of the XmlSource driver so that it
handles multiple field values correctly. Previously it was only
recording the first example, instead of creating a list. In addition
there was a bug that was causing multiple duplicate rows to be returned,
which was caused by the `npartitions` logic used by intake. There is
only one partition for the data since it is a single XML document that
is read in and processed.

The test for XmlSource was simplified to use requests_mock, and also to
look for values in the resulting DataFrame rather than comparing
DataFrames.

The desired behavior is for XML like:

```xml
<records>
    <record>
        <id>1</id>
        <title>Article 1</title>
        <name>Donna Haraway</name>
        <name>bell hooks</name>
    </record>
    <record>
        <id>2</id>
        <title>Article 2</title>
        <name>Elinor Ostrom</name>
        <name>Ivan Illich</name>
    </record>
</records>
```

and a metadata configuration like:

```python
metadata = {
    "record_selector": {
        "path": ".//record",
        "namespace": None
    },
    "fields": {
        "id": {
            "path": "id",
            "namespace": None
        },
        "title": {
            "path": "title",
            "namespace": None,
        },
        "name": {
            "path": "name",
            "namespace": None
        }
    }
}
```

To result in CSV that looks like:

```
id,title,name
1,Article 1,["Donna Harraway", "bell hooks"]
2,Article 2,["Elinor Ostrom", "Ivan Illich"]
```

Closes #243