Fix bin/get --limit
In 38df69fcbb5bfc2e33375db3babb723ffe478840 we stopped instantiating one
catalog per Collection and Provider in the constructor and instantiate a
new one each time the `catalog` property is used. This means that you
can't actually modify the one that will get used by intake. This isn't
usually an issue except for the `bin/get` utility with tries to set
`.record_limit` after the fact, but prior to running the harvest.