allow_expiration
Add an optional argument `allow_expiration` to the OaiXmlSource to allow
it to catch BadResumptionToken exceptions from Sickle. When caught a
warning is written to the log and the records that have been collected
so far are used.

By default `allow_expiration` is set to `False`, so that we know when we
are running into this problem with a provider, since catching the
exception results in incomplete data for a provider/collection.

Additional tests for resumtion token exceptions have been added. Tests
have also been refactored to use requests_mock, which simplifies them
significantly.

Closes #297