read_csv_with_lists()
When doing post-harvest tasks for QNL we need to make sure we convert
the lists that have been serialized as strings in the CSV back into
lists again, prior to trying to run merge_df. Otherwise these are
processed as strings, which results in lists where every member is a
single character!