Archive task (#371)
Add an archive task to the end of the ETL pipeline. The task will
archive the harvested data to an archive folder within the working
directory using a timestamped filename.

If the harvested data is empty (just a CSV header) then the archive is
not written. If the harvested data is identical (sha256 hash) of the
previous archive it also is not written. This is done to prevent the
archive directory from filling up with unnecessary files.

Closes #364