Adjust QNL post-harvest to use JSON
These are changes to get the post-harvest task for QNL to use the JSON
file that has been saved during harvest.

The dlme_airflow.utils.qnl.merge_records() function needed to be
adjusted to not turn the `id` column (the one column in the dataframe
that is a string and not a list) into a list where each element was a
character of the string. This was done by temporarily putting the id in
a list and then switching it back after the merge.

Fixes #392